# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BCR - Car Management API (OpenAPI 3.0)
  version: "1.0"
  description: "API to manage Binar Car Rental service website. This API is a part of fullstack development course individual project by Binar Academy. \n\n User is divided into three roles: superadmin, admin, member \n\n Superadmin default account details can be found on this project documentation. Superadmin account details can also be modified by following the instruction there."
  contact:
    name: sahat parulian
    email: sahatparulian85@gmail.com
    url: https://www.linkedin.com/in/sahat-parulian/
  license:
    name: MIT License
servers:
  - url: http://localhost:8000/api
  - url: http://localhost:8000/api/v1
...
paths:
  /register:
    post:
      tags:
        - auth
      summary: Register new member
      description: Register new account with member role
      operationId: registerMember
      security:
        - {}
      requestBody:
        description: Required fields [name, email, password]
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCred"

      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: create user successfully
                  data:
                    $ref: "#/components/schemas/User"
        "409":
          description: Conflict, email already registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
                    example: email already registered
        "422":
          description: Unprocessable Entity, check missing fields from requestBody
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissingFieldMsg"

        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultErrMsg"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    DefaultErrMsg:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Error message
...
    MissingFieldMsg:
      type: object
      properties:
        status:
          type: string
          example: failed
        message:
          type: string
          example: Missing required fields
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 4
        name:
          type: string
          example: yosua parulian
        email:
          type: string
          format: email
          example: yosuaparulian@gmail.com
        role:
          type: string
          enum:
            - member
            - admin
            - superadmin
          example: member
...
    RegisterCred:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          example: yosua parulian
        email:
          type: string
          format: email
          example: yosua@gmail.com
        password:
          type: string
          example: this_password
...
security:
  - bearerAuth: []